Added formatted input component.
When clicked, it acts like a normal plain text input. However, when the
input is blurred, it runs the input through an optional formatter
function, and then run through mathjax. Note: still need to do
sanitization because the user could input math formatting characters,
causing the whole thing to be formatted even when run through a plain
text formatter.