Implemented real-time fact set editor.
Added InjectorRef
Modified editor.component to use InjectorRef, so that child classes do
not have to pass the multiple injections it requires.
Added useful functions to the editor class such as createString and
beginCompoundOperation, so that subclasses do not have to interact with
the document model directly.
Fixed some Angular-related bugs in Editor.
Made a realtime version of the old fact set editor, which will be
removed shortly.