Added / fixed several different things, mostly used for <book-editor>.
Fixed bug where clicking the sign in button while GAPI is loading will
cause the sign in button to break.
Created <expanding-text-area> which is like <textarea> but expands to
fit its content. (It's still a bit buggy)
Created <list-item> which holds whatever content and adds a delete
button, which fires an event when clicked.
Added book editor, for taking notes on books, Currently, it only has
characters, but more types of objects are planned for the future.
Added <hide-box>, which is basically a collapsible div with a title.
Made a few cosmetic changes.