Added more stuff to <book-editor-component>
It now has events!
At least one name is required for all things that can be named (which is
characters and events, currently.)
Some cleaning up of the formatting, I think.
Some things can be selected by drop-down box now.
Added styling to drop-down boxes to make it more consistent with the
other widgets.
I'm going to work on tidying up the html for <book-editor-component> by
splitting it up into several child components, the current
implementation is a large bowl of spaghetti.