Implemented the showoff system.
Basically, hide boxes can have a 'path' of sorts, which is an array of
strings called 'showoffId'. Whenver the user focuses on a particular
hide box, a collaborative variable corresponding to that user is set to
the showoffId of the box they just focused on. Everyone will now be able
to see their icon rendered in the header of that box. Additionally, any
parent hide boxes (by showoffId, not by DOM hierarchy) will show the
user icon.

An example:

Box A (id: ['a'])        V
  Box 1 (id: ['a', '1']) V
    Content
  Box 2 (id: ['a', '2']) V
    Content
Box B (id: ['b'])        V
  Content
    
If the user clicked on box A, then their icon will be rendered in the
header of box A, and box A only. If the user clicked on box 1, then
their icon would be rendered in the headers of box A and box 1. If the
user then clicked on box B, their icon would only be rendered in the
header of box B.